Remove duplicate PrimaryValue fields from Elasticsearch models.
