+++ aspnetcore/src/api/Models/Elasticsearch/ElasticsearchFieldOfScience.cs
@@ -9,12 +9,10 @@ namespace api.Models.Elasticsearch
             NameFi = "";
             NameEn = "";
             NameSv = "";
-            PrimaryValue = null;
         }
 
         public string NameFi { get; set; }
         public string NameEn { get; set; }
         public string NameSv { get; set; }
-        public bool? PrimaryValue { get; set; }
     }
 }
