+++ aspnetcore/src/api/Models/Elasticsearch/ElasticsearchResearcherDescription.cs
@@ -7,12 +7,10 @@
             ResearchDescriptionFi = "";
             ResearchDescriptionEn = "";
             ResearchDescriptionSv = "";
-            PrimaryValue = null;
         }
 
         public string ResearchDescriptionFi { get; set; }
         public string ResearchDescriptionEn { get; set; }
         public string ResearchDescriptionSv { get; set; }
-        public bool? PrimaryValue { get; set; }
     }
 }
