+++ aspnetcore/src/api/Models/Elasticsearch/ElasticsearchTelephoneNumber.cs
@@ -7,10 +7,8 @@ namespace api.Models.Elasticsearch
         public ElasticsearchTelephoneNumber()
         {
             Value = "";
-            PrimaryValue = null;
         }
 
         public string Value { get; set; }
-        public bool? PrimaryValue { get; set; }
     }
 }
